refactor: Move rules to specific files
