+++ rules/no-svelte-store-in-api.js
@@ -1,0 +1,29 @@
+module.exports = {
+  meta: {
+    docs: {
+      description:
+        "Svelte stores should not be used in APIs since they are also utilized by workers.",
+      category: "Possible Errors",
+      recommended: false,
+    },
+    schema: [],
+  },
+  create(context) {
+    return {
+      ImportDeclaration(node) {
+        const filePath = context.getFilename();
+
+        const {
+          source: { value },
+        } = node;
+
+        if (filePath.includes("/api/") && value === "svelte/store") {
+          context.report({
+            node,
+            message: "Importing 'svelte/store' is not allowed in API modules.",
+          });
+        }
+      },
+    };
+  },
+};
